:art: Add functionality to parse prot.accession2taxid.FULL.gz :art: Move search_prot_accessions(...) method to NCBI class :art: Add checks for prot.accession2taxid.FULL.gz in NCBI class
